+++ tests/e2e/setup/setup.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"path"
 	"path/filepath"
-	"regexp"
 	"strconv"
 	"strings"
 	"time"
@@ -22,8 +21,6 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-	"github.com/ethereum/go-ethereum/common/hexutil"
-	"github.com/ethereum/go-ethereum/ethclient"
 	"github.com/ory/dockertest/v3"
 	"github.com/ory/dockertest/v3/docker"
 	bech32ibctypes "github.com/osmosis-labs/bech32-ibc/x/bech32ibc/types"
@@ -49,11 +46,9 @@ type E2ETestSuite struct {
 
 	tmpDirs             []string
 	Chain               *chain
-	EthClient           *ethclient.Client
 	gaiaRPC             *rpchttp.HTTP
 	DkrPool             *dockertest.Pool
 	DkrNet              *dockertest.Network
-	EthResource         *dockertest.Resource
 	GaiaResource        *dockertest.Resource
 	HermesResource      *dockertest.Resource
 	priceFeederResource *dockertest.Resource
@@ -84,31 +79,7 @@ func (s *E2ETestSuite) SetupSuite() {
 	s.DkrNet, err = s.DkrPool.CreateNetwork(fmt.Sprintf("%s-testnet", s.Chain.ID))
 	s.Require().NoError(err)
 
-	s.T().Logf("Ethereum and peggo are disable due to Ethereum PoS migration and PoW fork")
-	// var useGanache bool
-	// if str := os.Getenv("UMEE_E2E_USE_GANACHE"); len(str) > 0 {
-	// 	useGanache, err = strconv.ParseBool(str)
-	// 	s.Require().NoError(err)
-	// }
-
-	// The bootstrapping phase is as follows:
-	//
-	// 1. Initialize Umee validator nodes.
-	// 2. Launch an Ethereum container that mines.
-	// 3. Create and initialize Umee validator genesis files (setting delegate keys for validators).
-	// 4. Start Umee network.
-	// 5. Run an Oracle price feeder.
-	// 6. Create and run Gaia container(s).
-	// 7. Create and run IBC relayer (Hermes) containers.
-	// 8. Deploy the Gravity Bridge contract.
-	// 9. Create and start Peggo (orchestrator) containers.
-	s.initNodes()
-	// if useGanache {
-	// 	s.runGanacheContainer()
-	// } else {
-	// 	s.initEthereum()
-	// 	s.runEthContainer()
-	// }
+	s.initNodes() // init validator nodes
 	s.initGenesis()
 	s.initValidatorConfigs()
 	s.runValidators()
@@ -119,8 +90,6 @@ func (s *E2ETestSuite) SetupSuite() {
 	} else {
 		s.T().Log("running minimum network withut gaia,price-feeder and ibc-relayer")
 	}
-	// s.runContractDeployment()
-	// s.runOrchestrators()
 	s.initUmeeClient()
 }
 
@@ -136,7 +105,6 @@ func (s *E2ETestSuite) TearDownSuite() {
 
 	s.T().Log("tearing down e2e integration test suite...")
 
-	// s.Require().NoError(s.DkrPool.Purge(s.ethResource))
 	if !s.MinNetwork {
 		s.Require().NoError(s.DkrPool.Purge(s.GaiaResource))
 		s.Require().NoError(s.DkrPool.Purge(s.HermesResource))
@@ -146,10 +114,6 @@ func (s *E2ETestSuite) TearDownSuite() {
 	for _, vc := range s.ValResources {
 		s.Require().NoError(s.DkrPool.Purge(vc))
 	}
-
-	// for _, oc := range s.OrchResources {
-	// 	s.Require().NoError(s.DkrPool.Purge(oc))
-	// }
 
 	s.Require().NoError(s.DkrPool.RemoveNetwork(s.DkrNet))
 
@@ -161,7 +125,6 @@ func (s *E2ETestSuite) TearDownSuite() {
 
 func (s *E2ETestSuite) initNodes() {
 	s.Require().NoError(s.Chain.createAndInitValidators(s.cdc, 3))
-	s.Require().NoError(s.Chain.createAndInitOrchestrators(s.cdc, 3))
 	s.Require().NoError(s.Chain.createAndInitGaiaValidator(s.cdc))
 
 	// initialize a genesis file for the first validator
@@ -174,16 +137,6 @@ func (s *E2ETestSuite) initNodes() {
 		)
 	}
 
-	// add orchestrator accounts to genesis file
-	for _, orch := range s.Chain.Orchestrators {
-		orchAddr, err := orch.keyInfo.GetAddress()
-		s.Require().NoError(err)
-
-		s.Require().NoError(
-			addGenesisAccount(s.cdc, val0ConfigDir, "", InitBalanceStr, orchAddr),
-		)
-	}
-
 	// copy the genesis file to the remaining validators
 	for _, val := range s.Chain.Validators[1:] {
 		_, err := copyFile(
@@ -192,31 +145,6 @@ func (s *E2ETestSuite) initNodes() {
 		)
 		s.Require().NoError(err)
 	}
-}
-
-func (s *E2ETestSuite) initEthereum() {
-	// generate ethereum keys for validators add them to the ethereum genesis
-	ethGenesis := EthereumGenesis{
-		Difficulty: "0x400",
-		GasLimit:   "0xB71B00",
-		Config:     EthereumConfig{ChainID: EthChainID},
-		Alloc:      make(map[string]Allocation, len(s.Chain.Validators)+1),
-	}
-
-	alloc := Allocation{
-		Balance: "0x1337000000000000000000",
-	}
-
-	ethGenesis.Alloc["0xBf660843528035a5A4921534E156a27e64B231fE"] = alloc
-	for _, orch := range s.Chain.Orchestrators {
-		s.Require().NoError(orch.generateEthereumKey())
-		ethGenesis.Alloc[orch.EthereumKey.Address] = alloc
-	}
-	ethGenBz, err := json.MarshalIndent(ethGenesis, "", "  ")
-	s.Require().NoError(err)
-
-	// write out the genesis file
-	s.Require().NoError(writeFile(filepath.Join(s.Chain.configDir(), "eth_genesis.json"), ethGenBz))
 }
 
 func (s *E2ETestSuite) initGenesis() {
@@ -346,13 +274,7 @@ func (s *E2ETestSuite) initGenesis() {
 		}
 		s.Require().NoError(err)
 
-		orchAddr, err := s.Chain.Orchestrators[i].keyInfo.GetAddress()
-		s.Require().NoError(err)
-
-		delKeysMsg, err := val.buildDelegateKeysMsg(orchAddr, s.Chain.Orchestrators[i].EthereumKey.Address)
-		s.Require().NoError(err)
-
-		signedTx, err := val.signMsg(s.cdc, createValmsg, delKeysMsg)
+		signedTx, err := val.signMsg(s.cdc, createValmsg)
 		s.Require().NoError(err)
 
 		txRaw, err := s.cdc.MarshalJSON(signedTx)
@@ -424,147 +346,6 @@ func (s *E2ETestSuite) initValidatorConfigs() {
 
 		srvconfig.WriteConfigFile(appCfgPath, appConfig)
 	}
-}
-
-func (s *E2ETestSuite) runGanacheContainer() {
-	s.T().Log("starting Ganache container...")
-
-	tmpDir, err := os.MkdirTemp("", "umee-e2e-testnet-eth-")
-	s.Require().NoError(err)
-	s.tmpDirs = append(s.tmpDirs, tmpDir)
-
-	_, err = copyFile(
-		filepath.Join("./docker/", "ganache.Dockerfile"),
-		filepath.Join(tmpDir, "ganache.Dockerfile"),
-	)
-	s.Require().NoError(err)
-
-	entrypoint := []string{
-		"ganache-cli",
-		"-h",
-		"0.0.0.0",
-		"--networkId",
-		"15",
-	}
-
-	entrypoint = append(entrypoint, "--account", "0xb1bab011e03a9862664706fc3bbaa1b16651528e5f0e7fbfcbfdd8be302a13e7,0x3635C9ADC5DEA00000")
-	for _, orch := range s.Chain.Orchestrators {
-		s.Require().NoError(orch.generateEthereumKey())
-		entrypoint = append(entrypoint, "--account", orch.EthereumKey.PrivateKey+",0x3635C9ADC5DEA00000")
-	}
-
-	s.EthResource, err = s.DkrPool.BuildAndRunWithBuildOptions(
-		&dockertest.BuildOptions{
-			Dockerfile: "ganache.Dockerfile",
-			ContextDir: tmpDir,
-		},
-		&dockertest.RunOptions{
-			Name:         "ganache",
-			NetworkID:    s.DkrNet.Network.ID,
-			ExposedPorts: []string{"8545"},
-			PortBindings: map[docker.Port][]docker.PortBinding{
-				"8545/tcp": {{HostIP: "", HostPort: "8545"}},
-			},
-			Env:        []string{},
-			Entrypoint: entrypoint,
-		},
-		noRestart,
-	)
-	s.Require().NoError(err)
-
-	s.EthClient, err = ethclient.Dial(fmt.Sprintf("http://%s", s.EthResource.GetHostPort("8545/tcp")))
-	s.Require().NoError(err)
-
-	match := "Listening on 0.0.0.0:8545"
-
-	var (
-		outBuf bytes.Buffer
-		errBuf bytes.Buffer
-	)
-
-	// Wait for Ganache to start running.
-	s.Require().Eventually(
-		func() bool {
-			err := s.DkrPool.Client.Logs(
-				docker.LogsOptions{
-					Container:    s.EthResource.Container.ID,
-					OutputStream: &outBuf,
-					ErrorStream:  &errBuf,
-					Stdout:       true,
-					Stderr:       true,
-				},
-			)
-			if err != nil {
-				return false
-			}
-
-			return strings.Contains(outBuf.String(), match)
-		},
-		1*time.Minute,
-		5*time.Second,
-		"ganache node failed to start",
-	)
-	s.T().Logf("started Ganache container: %s", s.EthResource.Container.ID)
-}
-
-func (s *E2ETestSuite) runEthContainer() {
-	s.T().Log("starting Ethereum container...")
-	tmpDir, err := os.MkdirTemp("", "umee-e2e-testnet-eth-")
-	s.Require().NoError(err)
-	s.tmpDirs = append(s.tmpDirs, tmpDir)
-
-	_, err = copyFile(
-		filepath.Join(s.Chain.configDir(), "eth_genesis.json"),
-		filepath.Join(tmpDir, "eth_genesis.json"),
-	)
-	s.Require().NoError(err)
-
-	_, err = copyFile(
-		filepath.Join("./docker/", "eth.Dockerfile"),
-		filepath.Join(tmpDir, "eth.Dockerfile"),
-	)
-	s.Require().NoError(err)
-
-	s.EthResource, err = s.DkrPool.BuildAndRunWithBuildOptions(
-		&dockertest.BuildOptions{
-			Dockerfile: "eth.Dockerfile",
-			ContextDir: tmpDir,
-		},
-		&dockertest.RunOptions{
-			Name:      "ethereum",
-			NetworkID: s.DkrNet.Network.ID,
-			PortBindings: map[docker.Port][]docker.PortBinding{
-				"8545/tcp": {{HostIP: "", HostPort: "8545"}},
-			},
-			Env: []string{},
-		},
-		noRestart,
-	)
-	s.Require().NoError(err)
-
-	s.EthClient, err = ethclient.Dial(fmt.Sprintf("http://%s", s.EthResource.GetHostPort("8545/tcp")))
-	s.Require().NoError(err)
-
-	// Wait for the Ethereum node to start producing blocks; DAG completion takes
-	// about two minutes.
-	s.Require().Eventually(
-		func() bool {
-			ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
-			defer cancel()
-
-			height, err := s.EthClient.BlockNumber(ctx)
-			if err != nil {
-				return false
-			}
-
-			return height > 1
-		},
-		5*time.Minute,
-		10*time.Second,
-		"geth node failed to produce a block",
-	)
-
-	s.T().Logf("started Ethereum container: %s", s.EthResource.Container.ID)
 }
 
 func (s *E2ETestSuite) runValidators() {
@@ -791,177 +572,6 @@ func (s *E2ETestSuite) runIBCRelayer() {
 
 	// create the client, connection and channel between the Umee and Gaia chains
 	s.connectIBCChains()
-}
-
-func (s *E2ETestSuite) runContractDeployment() {
-	s.T().Log("starting Gravity Bridge contract deployer container...")
-
-	resource, err := s.DkrPool.RunWithOptions(
-		&dockertest.RunOptions{
-			Name:       "gravity-contract-deployer",
-			NetworkID:  s.DkrNet.Network.ID,
-			Repository: "umee-network/umeed-e2e",
-			// NOTE: container names are prefixed with '/'
-			Env: []string{"PEGGO_ETH_PK=" + EthMinerPK},
-			Entrypoint: []string{
-				"peggo",
-				"bridge",
-				"deploy-gravity",
-				"--eth-rpc",
-				fmt.Sprintf("http://%s:8545", s.EthResource.Container.Name[1:]),
-				"--cosmos-grpc",
-				fmt.Sprintf("tcp://%s:9090", s.ValResources[0].Container.Name[1:]),
-				"--tendermint-rpc",
-				fmt.Sprintf("http://%s:26657", s.ValResources[0].Container.Name[1:]),
-			},
-		},
-		noRestart,
-	)
-	s.Require().NoError(err)
-
-	s.T().Logf("started contract deployer: %s", resource.Container.ID)
-
-	// wait for the container to finish executing
-	container := resource.Container
-	for container.State.Running {
-		time.Sleep(10 * time.Second)
-
-		container, err = s.DkrPool.Client.InspectContainer(resource.Container.ID)
-		s.Require().NoError(err)
-	}
-
-	var (
-		outBuf bytes.Buffer
-		errBuf bytes.Buffer
-	)
-
-	s.Require().NoErrorf(s.DkrPool.Client.Logs(
-		docker.LogsOptions{
-			Container:    resource.Container.ID,
-			OutputStream: &outBuf,
-			ErrorStream:  &errBuf,
-			Stdout:       true,
-			Stderr:       true,
-		},
-	),
-		"failed to start contract deployer; stdout: %s, stderr: %s",
-		outBuf.String(), errBuf.String(),
-	)
-
-	re := regexp.MustCompile(`Address: (0x.+)`)
-	tokens := re.FindStringSubmatch(errBuf.String())
-	s.Require().Len(tokens, 2)
-
-	gravityContractAddr := tokens[1]
-	s.Require().NotEmpty(gravityContractAddr)
-
-	re = regexp.MustCompile(`Transaction: (0x.+)`)
-	tokens = re.FindStringSubmatch(errBuf.String())
-	s.Require().Len(tokens, 2)
-
-	txHash := tokens[1]
-	s.Require().NotEmpty(txHash)
-
-	s.Require().Eventually(
-		func() bool {
-			ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-			defer cancel()
-
-			if err := s.QueryEthTx(ctx, s.EthClient, txHash); err != nil {
-				return false
-			}
-
-			return true
-		},
-		time.Minute,
-		time.Second,
-		"failed to confirm Peggy contract deployment transaction",
-	)
-
-	s.Require().NoError(s.DkrPool.RemoveContainerByName(container.Name))
-
-	s.T().Logf("deployed Gravity Bridge contract: %s", gravityContractAddr)
-	s.GravityContractAddr = gravityContractAddr
-}
-
-func (s *E2ETestSuite) runOrchestrators() {
-	s.T().Log("starting orchestrator containers...")
-
-	s.OrchResources = make([]*dockertest.Resource, len(s.Chain.Validators))
-	for i, orch := range s.Chain.Orchestrators {
-		resource, err := s.DkrPool.RunWithOptions(
-			&dockertest.RunOptions{
-				Name:       s.Chain.Orchestrators[i].instanceName(),
-				NetworkID:  s.DkrNet.Network.ID,
-				Repository: "umee-network/umeed-e2e",
-				Env:        []string{"PEGGO_ETH_PK=" + orch.EthereumKey.PrivateKey},
-				// NOTE: container names are prefixed with '/'
-				Entrypoint: []string{
-					"peggo",
-					"orchestrator",
-					s.GravityContractAddr,
-					"--eth-rpc",
-					fmt.Sprintf("http://%s:8545", s.EthResource.Container.Name[1:]),
-					"--cosmos-chain-id",
-					s.Chain.ID,
-					"--cosmos-grpc",
-					fmt.Sprintf("tcp://%s:9090", s.ValResources[i].Container.Name[1:]),
-					"--tendermint-rpc",
-					fmt.Sprintf("http://%s:26657", s.ValResources[i].Container.Name[1:]),
-					"--cosmos-gas-prices",
-					minGasPrice,
-					"--cosmos-from",
-					orch.keyInfo.Name,
-					"--relay-batches=true",
-					"--valset-relay-mode=minimum",
-					"--profit-multiplier=0.0",
-					"--oracle-providers=mock",
-					"--relayer-loop-multiplier=1.0",
-					"--requester-loop-multiplier=1.0",
-					"--cosmos-pk",
-					hexutil.Encode(s.Chain.Orchestrators[i].PrivateKey.Bytes()),
-				},
-			},
-			noRestart,
-		)
-		s.Require().NoError(err)
-
-		s.OrchResources[i] = resource
-		s.T().Logf("started orchestrator container: %s", resource.Container.ID)
-	}
-
-	match := "oracle sent set of claims successfully"
-	for _, resource := range s.OrchResources {
-		s.T().Logf("waiting for orchestrator to be healthy: %s", resource.Container.ID)
-
-		var (
-			outBuf bytes.Buffer
-			errBuf bytes.Buffer
-		)
-
-		s.Require().Eventuallyf(
-			func() bool {
-				err := s.DkrPool.Client.Logs(
-					docker.LogsOptions{
-						Container:    resource.Container.ID,
-						OutputStream: &outBuf,
-						ErrorStream:  &errBuf,
-						Stdout:       true,
-						Stderr:       true,
-					},
-				)
-				if err != nil {
-					return false
-				}
-
-				return strings.Contains(errBuf.String(), match)
-			},
-			5*time.Minute,
-			time.Second,
-			"orchestrator %s not healthy",
-			resource.Container.ID,
-		)
-	}
 }
 
 func (s *E2ETestSuite) runPriceFeeder() {